requirements.txt
langchain
openai
tiktoken
faiss-cpu
gradio
unstructured
pdf2image
pdfminer-six
python-dotenv
azure-storage-blob
unstructured[doc]
chromadb